+++ src/aeolus/__init__.py
@@ -81,6 +81,7 @@ from .api import (
     get_current,
     get_source_info,
     list_sources,
+    summarize,
 )
 
 # Define what gets exported with "from aeolus import *"
@@ -99,5 +100,6 @@ __all__ = [
     "find_sites",
     "get_current",
     "get_source_info",
+    "summarize",
     "fetch",
 ]